+++ package.json
@@ -33,6 +33,7 @@
     "type-check": "tsc --noEmit",
     "dev": "npx tsc -p tsconfig.node.json && cp src/snarky.d.ts dist/node/snarky.d.ts",
     "make": "make -C ../../.. snarkyjs",
+    "bindings": "cd ../../.. && ./scripts/update-snarkyjs-bindings.sh && cd src/lib/snarkyjs",
     "build": "rimraf ./dist/node && npx tsc -p tsconfig.node.json && cp -r src/snarkyjs-bindings/compiled/node_bindings dist/node/_node_bindings && node src/build/buildNode.js && cp src/snarky.d.ts dist/node/snarky.d.ts",
     "build:test": "npx tsc -p tsconfig.test.json && cp src/snarky.d.ts dist/node/snarky.d.ts",
     "build:node": "npm run build",
