+++ src/snarky.js
@@ -1,6 +1,6 @@
-import { getSnarky, withThreadPool } from './snarky/wrapper.js';
-import snarkySpec from './snarky/snarky-class-spec.js';
-import { proxyClasses } from './snarky/proxy.js';
+import { getSnarky, withThreadPool } from './snarkyjs-bindings/js/wrapper.js';
+import snarkySpec from './snarkyjs-bindings/js/snarky-class-spec.js';
+import { proxyClasses } from './snarkyjs-bindings/js/proxy.js';
 
 export {
   Field,
