--- src/snarky/wrapper.d.ts
@@ -1,11 +1,0 @@
-import * as wasm from '../snarkyjs-bindings/compiled/node_bindings/plonk_wasm.js';
-
-export { WasmModule, getWasm, getSnarky, withThreadPool };
-
-type WasmModule = typeof wasm;
-
-declare function getWasm(): WasmModule;
-
-declare function getSnarky(): any;
-
-declare function withThreadPool<T>(run: () => Promise<T>): Promise<T>;
