--- src/snarky/wrapper.js
@@ -1,10 +1,0 @@
-import { wasm, withThreadPool } from './node/node-backend.js';
-import { default as snarky } from '../_node_bindings/snarky_js_node.bc.cjs';
-
-export { getSnarky, getWasm, withThreadPool };
-
-let getSnarky = () => snarky;
-
-function getWasm() {
-  return wasm;
-}
